Test base 2, 8, 10 only
Test only for bases 2, 8, 10 since a custom implementation is needed for other bases, and will be done in the Coin Jam code. I would prefer not to duplicate the code for the tests, or worse, assume that the future Coin Jam string to int converter is correct in the test.