Pass new version of 'Coin Jam' test (closes #1 )
The ``Generates valid jamcoins with correct test case numbers`` test was changed to force the generation of large integers. Coin Jam now uses bigints. Also, `string digit |> bigint.Parse` is used to convert a digit to a number instead of `bigint digit` because the latter results in a code (e.g. 'A' => 65).